pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.1",
    "logger>=1.4",
    "numpy>=2.2.5",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "tqdm>=4.67.1",
    "werkzeug>=3.1.3",
    "flask-login>=0.6.3",
]
